Fix: checkout button wasn't calling the correct function.
Update: reorganized some code.
