+++ droneyourfood/lib/Shopping/ShoppingScreen.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+import 'package:droneyourfood/Shopping/ShoppingList.dart';
+
+class ShoppingScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(
+          'Shopping Cart',
+        ),
+      ),
+      body: ShoppingListWidget(),
+    );
+  }
+}
